refactor(dict): speed up `TranslateWord` when building a table
`TranslateWord` is one of the most time-consuming methods to build a table. By replacing map and set with unordered_map and unordered_set for the involved classes: WeightMap, WordMap and ReverseLookupTable, the time it takes to build a dictionary of about 128w words can be saved about 5 seconds averagely (total time: ~20s -> ~15s). It's probably safe to do this since these three classes seem to have no requirement for order.

Also I replace `push_back` with `emplace_back` for the involved variables of vector type, the latter is a faster method to append a element to a vector.